server/requirements.txt
flask
Flask-SQLAlchemy
flask-socketio
Pillow
albumentations
opencv-python
numpy
bcrypt
pyjwt
matplotlib
tqdm